backend/requirements.txt
flask>=2.0.1
ultralytics>=8.0.196
opencv-python>=4.8.0.74
numpy>=1.24.3
pandas>=2.0.3
easyocr>=1.6.2
torch>=1.13.1
torchvision>=0.14.1
requests>=2.31.0
Pillow>=10.0.1
flask_cors>=0.10.1
flask-socketio
python-socketio
python-engineio
